Add CPP capabilty.  This does NOT intialise static classes, but C++ does work. We need to check in detail that, eg. new uses malloc.
